Move Windows VT input state management to stdin_handler_windows.rs
Move enable_vt_input, restore_vt_input, ORIGINAL_CONSOLE_MODE, and
related windows-sys imports from the shared stdin_handler.rs into
stdin_handler_windows.rs to keep platform-specific code in _windows
files per project convention.
